Migrate from vim-pathogen to vim-plug
Finally getting rid of the submodule "hell". This should make it easier
to maintain installed plugins. Although I'm still not sure if I need to
add the autoload script together with my dotfiles, but this is it for
now.

I also intend to do some cleanup on the vimrc file.