Update README to add vim-plug instructions
This means there is no need to have vim-plug versioned in the repository
anymore. It's better to keep it not versioned because otherwise we would
need to push a new change everytime we :PlugUpgrade.

I also removed some of the plugins I didn't use anymore and tried to
better configure vim-airline. Still lots to do on vimrc though.