Add missing Ruby PATHs
I had to add a small "hacky" solution to macOS' way of dealing with
paths in nested login shells (when using tmux, for example). On some
instances it would duplicate some entries in the PATH variable.

I also took the opportunity to remove old Android specific paths that
I'm not currently using.